OnlineSVM pre-version.
Working on all the splits without segfaults and updating the weights. Also the weights of each iteration is not stored yet.
Java code is used to split the dataset. (Although the problem of segfaults was with not correctly passing the chunked datasets to find_max_violators_all and
infer_latent_all functions.. found from later diagnosis). But the Java code to split datasets will be handy in the future, to create randomised sub-sets etc. So
keeping this.

Only line 6 of the write-up is called multiple times. So this is a pre-version of the OnlineSVM. Still need to store the weights, appropriately modify the
objective function etc.

+++ latentssvm_relext.mosek.dd/java/src/javaHelpers/FindMaxViolatorHelperAll.java
@@ -155,8 +155,8 @@ public class FindMaxViolatorHelperAll {
 		double simFracParam = Double.parseDouble(args[2]);
 		String datasetStatsFile = args[3];
 		double rho;
-		long datasetStartIdx = Long.parseLong(args[8]);
-		long chunkSz = Long.parseLong(args[9]);
+		int datasetStartIdx = Integer.parseInt(args[8]);
+		int chunkSz = Integer.parseInt(args[9]);
 		LabelWeights [] zWeights = Utils.initializeLabelWeights(currentParametersFile);
 		ArrayList<DataItem> dataset = Utils.populateDataset(datasetFile, datasetStartIdx, chunkSz);
